find words method now returns a hash of words
in the hash, each key is a row and the value for each key is an array of arrays. within each array are the 6 coordinates of the letters of an "uno" or "dos". we should be able to simply #count the number of arrays for each key to determine the number of words created on the board